01_ai_agents_first/02_openrouter/open_router/pyproject.toml
[project]
name = "open-router"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
]
